# openapi/product.yaml
openapi: 3.0.3
info:
  title: "Article Manager REST API"
  description: "API for managing articles."
  version: "1.0.0"
servers:
  - url: "http://localhost:8181/api"
...
paths:
...
  /public/promotion/{promotionId}:
    get:
      summary: "Get a promotion by ID (public)"
      description: "Retrieves a specific promotion by its ID without authentication"
      operationId: "getPublicPromotionById"
      tags:
        - "Promotion"
      parameters:
        - name: "promotionId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
          description: "The unique identifier of the promotion"
      responses:
        200:
          description: "Promotion found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionDto'
        404:
          description: "Promotion not found"
...
  /public/business/{businessId}/article/{articleId}:
    get:
      summary: "Get article by ID"
      description: "Retrieves an article by its ID."
      operationId: "getArticleById"
      tags:
        - "Article"
      parameters:
        - name: "businessId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "articleId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Article found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDto'
        404:
          description: "Article not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ArticleCategoryDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Unique identifier of the business type"
        nameEn:
          type: "string"
          description: "English name of the business type"
        descriptionEn:
          type: "string"
          description: "English description of the business type"
        nameFr:
          type: "string"
          description: "French name of the business type"
        descriptionFr:
          type: "string"
          description: "French description of the business type"
...
    PromotionStatusDto:
      type: "string"
      enum:
        - "ACTIVE"
        - "EXPIRED"
        - "SCHEDULED"
...
    PromotionDto:
      type: "object"
      required:
        - "id"
        - "articleId"
        - "businessId"
        - "discountPercentage"
        - "startDate"
        - "endDate"
        - "status"
      properties:
        id:
          type: "string"
          format: "uuid"
        article:
          $ref: "#/components/schemas/ArticleDto"
        businessId:
          type: "string"
          format: "uuid"
        discountPercentage:
          type: "number"
          format: "double"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        description:
          type: "string"
        status:
          $ref: '#/components/schemas/PromotionStatusDto'
        createdAt:
          type: "string"
          format: "date-time"
        originalPrice:
          $ref: '#/components/schemas/MoneyDto'
        productName:
          type: "string"
        productDescription:
          type: "string"
        productImages:
          type: "object"
          additionalProperties:
            type: "string"
        productCategory:
          $ref: "#/components/schemas/ArticleCategoryDTO"
        productQuantity:
          type: integer
        productSpecifications:
          type: "object"
          additionalProperties:
            type: "string"
...
    MoneyDto:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: decimal
          example: 49.99
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          example: "USD"
...
    TagDto:
      type: "object"
      required:
        - "id"
        - "title"
        - "businessId"
      properties:
        id:
          type: string
          format: uuid
        title:
          type: "string"
          description: "Title of the tag"
        description:
          type: "string"
          description: "Description of the tag"
        businessId:
          type: string
          format: uuid
          description: "Business identifier"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArticleStatusDto:
      type: "string"
      enum:
        - "DRAFT"
        - "PUBLISHED"
        - "UNPUBLISHED"
    ArticleDto:
      type: "object"
      required:
        - "id"
        - "name"
        - "description"
        - "category"
        - "price"
        - "quantity"
        - "specifications"
        - "createdAt"
        - "updatedAt"
        - "businessId"
        - "createdBy"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
          example: "Laptop Dell XPS 13"
        description:
          type: "string"
        category:
          $ref: "#/components/schemas/ArticleCategoryDTO"
        price:
          $ref: '#/components/schemas/MoneyDto'
        quantity:
          type: integer
          example: 50
        specifications:
          type: "object"
          additionalProperties:
            type: "string"
          description: "Product specifications as key-value pairs (e.g. size: XXL)"
          example:
            size: "XXL"
            color: "Black"
        images:
          type: "object"
          additionalProperties:
            type: "string"
            format: "string"
        createdAt:
          type: "string"
          format: "date-time"
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: "string"
          format: "date-time"
        businessId:
          type: "string"
          format: "uuid"
        createdBy:
          type: "string"
        status:
          $ref: '#/components/schemas/ArticleStatusDto'
        tags:
          type: "array"
          items:
            $ref: '#/components/schemas/TagDto'
...