# src/utils/api/product/index.ts
// This file is auto-generated by @hey-api/openapi-ts

export { Article, type Options, Promotion, Tag } from "./sdk.gen";
export {
  type AddTagsToArticleData,
  type AddTagsToArticleDto,
  type AddTagsToArticleErrors,
  type AddTagsToArticleResponse,
  type AddTagsToArticleResponses,
  type ArticleCategoryDto,
  type ArticleDto,
  type ArticleIdDto,
  type ArticleIdsRequestDto,
  type ArticlePageableResponseDto,
  ArticleSortFieldDto,
  ArticleStatusDto,
  type ClientOptions,
  type CountAllPublishedArticlesData,
  type CountAllPublishedArticlesResponse,
  type CountAllPublishedArticlesResponses,
  type CountArticlesByBusinessIdData,
  type CountArticlesByBusinessIdErrors,
  type CountArticlesByBusinessIdResponse,
  type CountArticlesByBusinessIdResponses,
  type CountArticlesByUserData,
  type CountArticlesByUserErrors,
  type CountArticlesByUserResponse,
  type CountArticlesByUserResponses,
  type CountByUserQueryDto,
  type CountByUserRequestDto,
  type CountByUserResponseDto,
  type CountByUserResultDto,
  type CountResponseDto,
  type CreateArticleData,
  type CreateArticleDto,
  type CreateArticleErrors,
  type CreateArticleResponse,
  type CreateArticleResponses,
  type CreatePromotionData,
  type CreatePromotionDto,
  type CreatePromotionErrors,
  type CreatePromotionResponse,
  type CreatePromotionResponses,
  type CreateTagData,
  type CreateTagDto,
  type CreateTagErrors,
  type CreateTagResponse,
  type CreateTagResponses,
  type DeleteArticleByIdData,
  type DeleteArticleByIdErrors,
  type DeleteArticleByIdResponse,
  type DeleteArticleByIdResponses,
  type DeletePromotionData,
  type DeletePromotionErrors,
  type DeletePromotionResponse,
  type DeletePromotionResponses,
  type GetActivePromotionsData,
  type GetActivePromotionsResponse,
  type GetActivePromotionsResponses,
  type GetAllArticleCategoriesData,
  type GetAllArticleCategoriesResponse,
  type GetAllArticleCategoriesResponses,
  type GetAllArticlesByBusinessIdData,
  type GetAllArticlesByBusinessIdErrors,
  type GetAllArticlesByBusinessIdForBusinessUserData,
  type GetAllArticlesByBusinessIdForBusinessUserErrors,
  type GetAllArticlesByBusinessIdForBusinessUserResponse,
  type GetAllArticlesByBusinessIdForBusinessUserResponses,
  type GetAllArticlesByBusinessIdResponse,
  type GetAllArticlesByBusinessIdResponses,
  type GetAllArticlesData,
  type GetAllArticlesResponse,
  type GetAllArticlesResponses,
  type GetAllPromotionsData,
  type GetAllPromotionsResponse,
  type GetAllPromotionsResponses,
  type GetAllTagsByBusinessIdData,
  type GetAllTagsByBusinessIdErrors,
  type GetAllTagsByBusinessIdResponse,
  type GetAllTagsByBusinessIdResponses,
  type GetArticleByIdData,
  type GetArticleByIdErrors,
  type GetArticleByIdResponse,
  type GetArticleByIdResponses,
  type GetArticlesByIdsData,
  type GetArticlesByIdsErrors,
  type GetArticlesByIdsResponse,
  type GetArticlesByIdsResponses,
  type GetPromotionByIdData,
  type GetPromotionByIdErrors,
  type GetPromotionByIdResponse,
  type GetPromotionByIdResponses,
  type GetPromotionsByBusinessIdData,
  type GetPromotionsByBusinessIdErrors,
  type GetPromotionsByBusinessIdResponse,
  type GetPromotionsByBusinessIdResponses,
  type GetPublicPromotionByIdData,
  type GetPublicPromotionByIdErrors,
  type GetPublicPromotionByIdResponse,
  type GetPublicPromotionByIdResponses,
  type MoneyDto,
  type PageableResponsePageDto,
  type PromotionDto,
  type PromotionIdDto,
  type PromotionPageableResponseDto,
  PromotionSortFieldDto,
  PromotionStatusDto,
  type PublishArticleData,
  type PublishArticleErrors,
  type PublishArticleResponse,
  type PublishArticleResponses,
  SortDirectionDto,
  type TagDto,
  type TagIdDto,
  type TagPageableResponseDto,
  type UnpublishAllArticlesData,
  type UnpublishAllArticlesErrors,
  type UnpublishAllArticlesResponse,
  type UnpublishAllArticlesResponses,
  type UnpublishArticleData,
  type UnpublishArticleErrors,
  type UnpublishArticleResponse,
  type UnpublishArticleResponses,
  type UpdateArticleByIdData,
  type UpdateArticleByIdErrors,
  type UpdateArticleByIdResponse,
  type UpdateArticleByIdResponses,
  type UpdateArticleDto,
  type UpdatePromotionData,
  type UpdatePromotionDto,
  type UpdatePromotionErrors,
  type UpdatePromotionResponse,
  type UpdatePromotionResponses,
  type VerifyArticleOwnershipData,
  type VerifyArticleOwnershipResponse,
  type VerifyArticleOwnershipResponses,
  type VerifyOwnershipRequestDto,
  type VerifyOwnershipResponseDto,
} from "./types.gen";